feat(sui-ci): move from index the command
